feat: add server settings
